requirements.txt
fastapi
uvicorn
requests
pydantic
sqlalchemy
python-multipart
python-dotenv
